requirements.txt
python-telegram-bot==12.7
SQLAlchemy==1.3.17